+++ src/bp/core/api.ts
@@ -1,5 +1,6 @@
 import * as sdk from 'botpress/sdk'
-import { EventEngine } from 'core/events/event-engine'
+import { EventEngine, EventRepository } from 'core/events'
+import { KeyValueStore } from 'core/kvs'
 import { WellKnownFlags } from 'core/sdk/enums'
 import { inject, injectable } from 'inversify'
 import Knex from 'knex'
@@ -13,12 +14,11 @@ import modelIdService from 'nlu-core/model-id-service'
 import { container } from './app.inversify'
 import { ConfigProvider } from './config/config-loader'
 import Database from './database'
-import { KeyValueStore } from './kvs'
 import { LoggerProvider } from './logger'
 import { getMessageSignature } from './misc/security'
 import { renderRecursive } from './misc/templating'
 import { ModuleLoader } from './module-loader'
-import { EventRepository, UserRepository } from './repositories'
+import { UserRepository } from './repositories'
 import { Event, RealTimePayload } from './sdk/impl'
 import HTTPServer from './server'
 import { GhostService } from './services'
