+++ src/bp/core/botpress.ts
@@ -1,7 +1,7 @@
 import * as sdk from 'botpress/sdk'
 import lang from 'common/lang'
-import { addStepToEvent, EventCollector, StepScopes, StepStatus } from 'core/events/event-collector'
-import { EventEngine } from 'core/events/event-engine'
+import { buildUserKey, converseApiEvents } from 'core/converse'
+import { addStepToEvent, EventCollector, StepScopes, StepStatus, EventEngine } from 'core/events'
 import { LoggerDbPersister, LoggerFilePersister, LoggerProvider, LogsJanitor } from 'core/logger'
 import { copyDir } from 'core/misc/pkg-fs'
 import { WrapErrorsWith } from 'errors'
@@ -34,7 +34,6 @@ import AuthService from './services/auth/auth-service'
 import { BotMonitoringService } from './services/bot-monitoring-service'
 import { BotService } from './services/bot-service'
 import { CMSService } from './services/cms'
-import { buildUserKey, converseApiEvents } from './services/converse'
 import { DecisionEngine } from './services/dialog/decision-engine'
 import { DialogEngine } from './services/dialog/dialog-engine'
 import { DialogJanitor } from './services/dialog/janitor'
