+++ src/bp/core/collaborators/repositories/workspace_invite_codes.ts
@@ -1,8 +1,7 @@
+import Database from 'core/database'
+import { TYPES } from 'core/types'
 import { inject, injectable } from 'inversify'
 import _ from 'lodash'
-
-import Database from '../database'
-import { TYPES } from '../types'
 
 export interface WorkspaceInviteCode {
   workspaceId: string
