+++ src/bp/core/converse/converse-router.ts
@@ -1,14 +1,14 @@
 import { Logger } from 'botpress/sdk'
 import { StandardError } from 'common/http'
+import { ConverseService } from 'core/converse'
 import HTTPServer from 'core/server'
 import AuthService, { TOKEN_AUDIENCE } from 'core/services/auth/auth-service'
-import { ConverseService } from 'core/services/converse'
 import { RequestHandler, Router } from 'express'
 import joi from 'joi'
 import _ from 'lodash'
 
-import { CustomRouter } from '../customRouter'
-import { checkTokenHeader } from '../util'
+import { CustomRouter } from '../routers/customRouter'
+import { checkTokenHeader } from '../routers/util'
 
 // this schema ensures a non breaking api signature (> 11.5)
 // see https://botpress.com/docs/build/channels/#usage-public-api
