+++ src/bp/core/database/tables/bot-specific/index.ts
@@ -1,9 +1,6 @@
 export * from './dialog_sessions'
-export * from 'core/logger/logs-table'
 export * from './ghost_content'
 export * from './notifications'
-export * from 'core/events/event-table'
-export * from 'core/kvs/kvs-table'
 export * from './conversations'
 export * from './messages'
 export * from './tasks'
