+++ src/bp/core/database/tables/server-wide/index.ts
@@ -1,7 +1,5 @@
 export * from './channel_users'
 export * from './metadata'
 export * from './data_retention'
-export * from 'core/collaborators/tables/workspace_users'
-export * from './workspace_invite_codes'
 export * from './telemetry'
 export * from './migrations'
