+++ src/bp/core/events/middleware-chain.test.ts
@@ -4,7 +4,7 @@ import { createSpyObject, MockObject } from 'core/misc/utils'
 import 'jest-extended'
 import 'reflect-metadata'
 
-import MiddlewareChain from './middleware-chain'
+import { MiddlewareChain } from './middleware-chain'
 
 const base = { description: 'test', order: 1, direction: <EventDirection>'incoming' }
 describe('Middleware', () => {
