+++ src/bp/core/events/middleware-chain.ts
@@ -1,7 +1,8 @@
 import * as sdk from 'botpress/sdk'
-import { addStepToEvent, StepScopes, StepStatus } from 'core/events/event-collector'
 import _ from 'lodash'
 import ms from 'ms'
+
+import { addStepToEvent, StepScopes, StepStatus } from '.'
 
 interface MiddlewareChainOptions {
   timeoutInMs: number
@@ -11,7 +12,7 @@ const defaultOptions = {
   timeoutInMs: ms('2s')
 }
 
-class MiddlewareChain {
+export class MiddlewareChain {
   private stack: { mw: sdk.IO.MiddlewareHandler; name: string }[] = []
 
   constructor(private options: MiddlewareChainOptions = defaultOptions) {
@@ -49,5 +50,3 @@ class MiddlewareChain {
     }
   }
 }
-
-export default MiddlewareChain
