+++ src/bp/core/logger/index.ts
@@ -1,6 +1,7 @@
-export * from './logs-service'
 export * from './logs-janitor'
 export * from './logs-repository'
+export * from './logs-service'
+export * from './logs-table'
 export * from './persister/console-logger'
 export * from './persister/db-persister'
 export * from './persister/file-persister'
