+++ src/bp/core/logger/persister/console-logger.ts
@@ -2,7 +2,7 @@ import { AxiosError } from 'axios'
 import { IO, Logger, LoggerEntry, LoggerLevel, LoggerListener, LogLevel } from 'botpress/sdk'
 import chalk from 'chalk'
 import { Metric } from 'common/monitoring'
-import { addLogToEvent } from 'core/events/event-collector'
+import { addLogToEvent } from 'core/events'
 import { IDisposable } from 'core/misc/disposable'
 import { BotService } from 'core/services/bot-service'
 import { incrementMetric } from 'core/services/monitoring'
