+++ src/bp/core/server.ts
@@ -6,6 +6,7 @@ import { RequestWithUser } from 'common/typings'
 import compression from 'compression'
 import cookieParser from 'cookie-parser'
 import session from 'cookie-session'
+import { ConverseRouter, ConverseService } from 'core/converse'
 import { LogsService, LogsRepository } from 'core/logger'
 import cors from 'cors'
 import errorHandler from 'errorhandler'
@@ -35,7 +36,6 @@ import { ModuleLoader } from './module-loader'
 import { TelemetryRepository } from './repositories/telemetry'
 import { AdminRouter, AuthRouter, BotsRouter, MediaRouter, ModulesRouter } from './routers'
 import { ContentRouter } from './routers/bots/content'
-import { ConverseRouter } from './routers/bots/converse'
 import { HintsRouter } from './routers/bots/hints'
 import { NLURouter } from './routers/bots/nlu'
 import { isDisabled } from './routers/conditionalMiddleware'
@@ -53,7 +53,7 @@ import AuthService, { EXTERNAL_AUTH_HEADER, SERVER_USER, TOKEN_AUDIENCE } from '
 import { generateUserToken } from './services/auth/util'
 import { BotService } from './services/bot-service'
 import { CMSService } from './services/cms'
-import { ConverseService } from './services/converse'
+
 import { FlowService } from './services/dialog/flow/service'
 import { SkillService } from './services/dialog/skill/service'
 import { HintsService } from './services/hints'
