+++ src/bp/core/services/action/action-service.ts
@@ -3,7 +3,7 @@ import { IO, Logger } from 'botpress/sdk'
 import { extractEventCommonArgs } from 'common/action'
 import { ObjectCache } from 'common/object-cache'
 import { ActionScope, ActionServer, LocalActionDefinition } from 'common/typings'
-import { addErrorToEvent, addStepToEvent, StepScopes, StepStatus } from 'core/events/event-collector'
+import { addErrorToEvent, addStepToEvent, StepScopes, StepStatus } from 'core/events'
 import { UntrustedSandbox } from 'core/misc/code-sandbox'
 import { printObject } from 'core/misc/print'
 import { TasksRepository } from 'core/repositories/tasks'
