+++ src/bp/core/services/auth/auth-service.ts
@@ -5,7 +5,7 @@ import { StrategyUserTable } from 'core/collaborators/tables/strategy_users'
 import { AuthStrategy } from 'core/config/botpress.config'
 import { ConfigProvider } from 'core/config/config-loader'
 import Database from 'core/database'
-import { EventEngine } from 'core/events/event-engine'
+import { EventEngine } from 'core/events'
 import { KeyValueStore } from 'core/kvs'
 import { getMessageSignature } from 'core/misc/security'
 import { ModuleLoader } from 'core/module-loader'
