+++ src/bp/core/services/messaging/messages.ts
@@ -1,5 +1,6 @@
 import * as sdk from 'botpress/sdk'
-import { EventEngine } from 'core/events/event-engine'
+import { EventEngine } from 'core/events'
+import { KeyValueStore } from 'core/kvs'
 import { MessageRepository } from 'core/repositories/messages'
 import { IOEvent } from 'core/sdk/impl'
 import { inject, injectable, postConstruct } from 'inversify'
@@ -7,7 +8,6 @@ import { AppLifecycle, AppLifecycleEvents } from 'lifecycle'
 import LRU from 'lru-cache'
 import ms from 'ms'
 
-import { KeyValueStore } from '../../kvs'
 import { TYPES } from '../../types'
 import { JobService } from '../job-service'
 import { ConversationService, ScopedConversationService } from './conversations'
