+++ src/bp/core/services/services.inversify.ts
@@ -1,8 +1,7 @@
 import { IO } from 'botpress/sdk'
 import LicensingService from 'common/licensing-service'
-import { EventEngine } from 'core/events/event-engine'
-import { Queue } from 'core/events/queue'
-import MemoryQueue from 'core/events/queue/memory-queue'
+import { ConverseService } from 'core/converse'
+import { EventEngine, Queue, MemoryQueue } from 'core/events'
 import { KeyValueStore } from 'core/kvs'
 import { LogsJanitor, LogsService } from 'core/logger'
 import { DialogContainerModule } from 'core/services/dialog/dialog.inversify'
@@ -19,7 +18,6 @@ import AuthService from './auth/auth-service'
 import { BotMonitoringService } from './bot-monitoring-service'
 import { BotService } from './bot-service'
 import { CMSService } from './cms'
-import { ConverseService } from './converse'
 import { SkillService } from './dialog/skill/service'
 import { GhostContainerModule } from './ghost/ghost.inversify'
 import { HintsService } from './hints'
