+++ src/bp/core/services/workspace-service.ts
@@ -10,10 +10,13 @@ import {
 } from 'botpress/sdk'
 import { CHAT_USER_ROLE, defaultPipelines, defaultWorkspace } from 'common/defaults'
 import { AuthRole, CreateWorkspace, Pipeline, Workspace } from 'common/typings'
-import { StrategyUsersRepository } from 'core/collaborators/repositories/strategy_users'
-import { WorkspaceUsersRepository } from 'core/collaborators/repositories/workspace_users'
+import {
+  StrategyUsersRepository,
+  WorkspaceUsersRepository,
+  WorkspaceInviteCode,
+  WorkspaceInviteCodesRepository
+} from 'core/collaborators'
 import { ConfigProvider } from 'core/config/config-loader'
-import { WorkspaceInviteCode, WorkspaceInviteCodesRepository } from 'core/repositories'
 import { ConflictError, NotFoundError } from 'core/routers/errors'
 import { inject, injectable, tagged } from 'inversify'
 import _ from 'lodash'
